+++ src/components/OpenEbooksLanding.tsx
@@ -4,7 +4,6 @@ import * as React from "react";
 import { H2, Text } from "./Text";
 import Button, { NavButton } from "./Button";
 import Stack from "./Stack";
-import { ClientBasicMethod, ClientCleverMethod, OPDS1 } from "interfaces";
 import useLibraryContext from "./context/LibraryContext";
 import useUser from "./context/UserContext";
 import SignOut from "./SignOut";
@@ -16,8 +15,7 @@ import Footer from "components/Footer";
 import GlobalStyles from "components/GlobalStyles";
 import { ErrorBoundary } from "components/ErrorBoundary";
 import { APP_CONFIG } from "utils/env";
-import { AppSetupError } from "errors";
-import AuthButton from "auth/AuthButton";
+import OpenEbooksLoginPicker from "auth/OpenEbooksLoginPicker";
 
 type PopularBook = { alt: string; imgHref: string };
 
@@ -52,21 +50,6 @@ const LandingPage: NextPage<AppProps> = ({ library, error }) => {
 };
 
 export const OpenEbooksLandingComponent = () => {
-  const { authMethods } = useLibraryContext();
-
-  const cleverMethod = authMethods.find(
-    method => method.type === OPDS1.CleverAuthType
-  ) as ClientCleverMethod | undefined;
-
-  const basicMethod = authMethods.find(
-    method => method.type === OPDS1.BasicAuthType
-  ) as ClientBasicMethod | undefined;
-
-  if (!cleverMethod || !basicMethod)
-    throw new AppSetupError(
-      "Application is missing either Clever or Basic Auth methods"
-    );
-
   return (
     <div
       sx={{
@@ -100,77 +83,7 @@ export const OpenEbooksLandingComponent = () => {
           </Text>
         </div>
       </div>
-      <div
-        id="loginRegion"
-        sx={{
-          backgroundColor: "ui.gray.extraLight"
-        }}
-      >
-        <div
-          sx={{
-            maxWidth: 1100,
-            mx: "auto",
-            display: "flex",
-            textAlign: ["center", "center", "left"],
-            flexWrap: ["wrap", "wrap", "nowrap"]
-          }}
-        >
-          <Stack
-            direction="column"
-            sx={{
-              mx: [3, 5],
-              my: 4,
-              justifyContent: "space-between",
-              flexWrap: ["wrap", "nowrap"]
-            }}
-          >
-            <img
-              sx={{ alignSelf: "center" }}
-              alt="Clever Logo"
-              src={"/img/CleverLogo.png"}
-            />
-            <Text>
-              Clever is the platform that powers technology in the classroom.
-              Today, one in three innovative K-12 schools in the U.S. trust
-              Clever to secure their student data as they adopt learning apps in
-              the classroom.
-            </Text>
-            <div>
-              <AuthButton
-                sx={{ mx: ["auto", "auto", 0] }}
-                method={cleverMethod}
-              />
-            </div>
-          </Stack>
-          <Stack
-            direction="column"
-            sx={{
-              mx: [3, 5],
-              my: 4,
-              justifyContent: "space-between",
-              flexWrap: ["wrap", "nowrap"]
-            }}
-          >
-            <img
-              sx={{ alignSelf: "center" }}
-              alt="FirstBook Logo"
-              src={"/img/FirstBookLogo.png"}
-            />
-            <Text>
-              First Book is a nonprofit organization that provides access to
-              high quality, brand new books and educational resources - for free
-              and at low cost - to schools and programs serving children in
-              need.
-            </Text>
-            <div>
-              <AuthButton
-                sx={{ mx: ["auto", "auto", 0] }}
-                method={basicMethod}
-              />
-            </div>
-          </Stack>
-        </div>
-      </div>
+      <OpenEbooksLoginPicker />
       <div
         sx={{
           backgroundColor: "brand.primary"